Issue 48751: LKSM/LKB: Respect sort order of sources or samples for lineage sample creation
