+++ packages/components/src/internal/components/picklist/actions.ts
@@ -17,11 +17,11 @@ import { PICKLIST_KEY } from '../../app/constants';
 import { isProductProjectsEnabled } from '../../app/utils';
 
 import { SCHEMAS } from '../../schemas';
-import { OperationConfirmationData } from '../entities/models';
 import { caseInsensitive } from '../../util/utils';
 
 import { Picklist, PICKLIST_KEY_COLUMN, PICKLIST_SAMPLE_ID_COLUMN } from './models';
 import { PRIVATE_PICKLIST_CATEGORY, PUBLIC_PICKLIST_CATEGORY } from './constants';
+import {getOrderedSelectedMappedKeys} from "../entities/actions";
 
 export function getPicklistsForInsert(): Promise<Picklist[]> {
     return new Promise((resolve, reject) => {
@@ -212,96 +212,6 @@ export function getOrderedSelectedPicklistSamples(queryModel: QueryModel, saveSn
         queryParameters,
         viewName
     );
-}
-
-interface RemappedKeyValues {
-    mapFromValues: any[];
-    mapToValues: any[];
-}
-
-/**
- * Get the ordered remapped key values from a QueryModel based on grid's current selection.
- * For example, picklist grid has a "ID" PK column and a "SampleId" FK column.
- * This function can be used to get the SampleIds for the currently selected IDs, in the order that respect current grid
- * filter/sort
- * @param fromColumn Key column for the current grid
- * @param toColumn Key column for the FK field, can be empty.
- * @param selectedIds
- * @param queryModel
- */
-export function getOrderedSelectedMappedKeysFromQueryModel(
-    fromColumn: string,
-    toColumn: string,
-    queryModel: QueryModel,
-    selectedIds?: string[]
-): Promise<RemappedKeyValues> {
-    const { schemaName, queryName, queryParameters, sortString, viewName, selections } = queryModel;
-    return getOrderedSelectedMappedKeys(
-        fromColumn,
-        undefined,
-        schemaName,
-        queryName,
-        selectedIds ?? Array.from(selections),
-        sortString,
-        queryParameters,
-        viewName
-    );
-}
-
-function getOrderedSelectedMappedKeys(
-    fromColumn: string,
-    toColumn: string,
-    schemaName: string,
-    queryName: string,
-    selections: string[],
-    sortString?: string,
-    queryParameters?: Record<string, any>,
-    viewName?: string
-): Promise<RemappedKeyValues> {
-    return new Promise((resolve, reject) => {
-        getSelectedData(
-            schemaName,
-            queryName,
-            Array.from(selections),
-            toColumn ? [fromColumn, toColumn].join(',') : fromColumn,
-            sortString,
-            queryParameters,
-            viewName,
-            fromColumn
-        )
-            .then(response => {
-                const { data, dataIds } = response;
-                const values = [];
-                data.forEach(row => {
-                    const rowData = row.toJS();
-                    const from = caseInsensitive(rowData, fromColumn)?.value;
-                    const to = toColumn ? caseInsensitive(rowData, toColumn)?.value : null;
-                    const orderNum = dataIds.indexOf(from + '');
-                    values.push({
-                        from,
-                        to,
-                        orderNum,
-                    });
-                });
-
-                const mapFromValues = [];
-                const mapToValues = [];
-                values.sort((a, b) => a.orderNum - b.orderNum);
-                values.forEach(value => {
-                    mapToValues.push(value.to);
-                    mapFromValues.push(value.from);
-                });
-
-                resolve({
-                    mapToValues,
-                    mapFromValues,
-                });
-            })
-            .catch(reason => {
-                console.error(reason);
-                reject(reason);
-            });
-    });
 }
 
 export function getSelectedPicklistSamples(
